Cargo Clippy
Summary of fixes:

Fixed compilation errors in csv-explorer-storage, csv-ethereum, and csv-bitcoin
Auto-fixed warnings with cargo clippy --fix
Manually fixed 25+ clippy warning types across multiple crates
Added #[allow] attributes for intentional code patterns (too_many_arguments, missing_docs, dead_code, type_complexity, unexpected_cfgs, deprecated)
Fixed csv-wallet warnings (redundant_field_names, redundant_locals, unnecessary_filter_map, unnecessary_sort_by, vec_init_then_push, field_reassign_with_default)
Fixed example compilation errors (subscriptions, gaming, performance)
Fixed csv-cli compilation errors and warnings (module imports, doc comments, useless_vec)

+++ csv-bitcoin/src/tx_builder.rs
@@ -382,7 +382,7 @@ mod tests {
 
         assert!(result.fee_sat > 0);
         assert_eq!(result.input_value_sat, 1_000_000);
-        assert!(result.raw_tx.len() >= result.tx.vsize() as usize);
+        assert!(result.raw_tx.len() >= result.tx.vsize());
         assert_eq!(
             result.tapret_output.amount_sat,
             result.input_value_sat - result.fee_sat
